scripts/requirements.txt
absl-py==2.1.0
accelerate==0.34.2
addict==2.4.0
annotated-types==0.7.0
apex @ git+https://github.com/NVIDIA/apex@05091d498d21058a0fe736b828c43431d4f0dda2
astunparse==1.6.3
async-timeout==4.0.3
attrs==19.3.0
audioread==3.0.1
av==13.1.0
bidict==0.23.1
byted-dataloader==0.4.4
byted-iceberg==0.2.112
byted-torch @ http://luban-source.byted.org/repository/scm/arnold.benchmark.pytorch2_1.0.0.22.tar.gz#sha256=b7d9f84093f8f734b2b5fa312ae024c3d9e6bbbe204a9e835c467416d833de6b
byted-torch-monitor @ https://d.scm.byted.org/api/v2/download/lab.pytorch.torch_monitor_1.0.0.6.tar.gz#sha256=333bc7b5f093bb449cad019a5351702f7497fc7b5e4a3f1c82f6567e607c3e1e
byted-wandb==0.13.74
bytedance-context==0.7.1
bytedance-metrics==0.5.1
bytedance.modelhub==0.0.77
bytedance.servicediscovery==0.1.2
bytedbackgrounds==0.0.6
byteddatabus==1.0.6
byteddps==0.1.2
bytedenv==0.6.2
bytedeuler==0.42.4
bytedevent==1.5.0
bytedlego-ops @ https://luban-source.byted.org/repository/scm/data.aml.lego_ops_th21_cu121_cudnn890_abi0_sdist_1.0.0.13.tar.gz#sha256=068f4d253b7d185feda1ff9e10398579617667e41f352ee808457d02e0827875
bytedlogger==0.15.1
bytedlogid==0.2.1
bytedmemfd==0.2
bytedmetrics==0.10.2
bytedmoxing @ https://luban-source.byted.org/repository/scm/search.nlp.moxing_py_1.0.0.21.tar.gz#sha256=10a8d5ad88b1f12dcecd3350d1a7f414f6af966734a7b155c389baf5d5c1872b
bytedpymongo==2.0.5
bytedrh2==1.18.9a15
bytedservicediscovery==0.17.4
bytedtcc==1.4.2
bytedtitan @ http://luban-source.byted.org/repository/scm/data.aml.titan_1.0.0.151.tar.gz#sha256=69c69518832ba5931d9037c7d59167f3bf01104bc4c9f4cc90c9dd8567f754bb
bytedtos==1.1.6
bytedtrace==0.3.0
bytedztijwthelper==0.0.23
bytedztispiffe==0.0.14
cachetools==5.3.3
certifi==2024.6.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
cmake==3.29.3
contourpy==1.2.1
cryptography==42.0.8
cycler==0.12.1
dbus-python==1.2.16
decorator==5.1.1
decord==0.6.0
deepspeed==0.15.1
Deprecated==1.2.14
devscripts==2.21.3+deb11u1
dill==0.3.8
distlib==0.3.8
distro-info==1.0+deb11u1
dnspython==2.6.1
docker-pycreds==0.4.0
docstring_parser==0.16
einops==0.8.0
filelock==3.14.0
flash-attn==1.0.3.post0
flatbuffers==24.3.25
fonttools==4.53.0
fsspec==2024.9.0
fvcore==0.1.5.post20221221
gast==0.5.4
gevent==22.10.2
gitdb==4.0.11
GitPython==3.1.43
google-auth==2.29.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
greenlet==3.0.3
grpcio==1.64.0
gunicorn==20.1.0
h11==0.14.0
h5py==3.11.0
hjson==3.1.0
huggingface-hub==0.25.2
idna==3.7
importlib_metadata==7.1.0
importlib_resources==6.4.0
iniconfig==2.0.0
iopath==0.1.10
iotop==0.6
ipaddress==1.0.23
iso8601==1.0.0
Jinja2==3.1.4
joblib==1.4.2
jsonargparse==4.14.0
jsonpath-ng==1.5.3
keras==2.8.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.5
lazy_loader==0.4
libclang==18.1.1
librosa==0.10.2.post1
llvmlite==0.42.0
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.0
mdurl==0.1.2
mmhash3==3.0.1
mpi4py==3.1.6
mpmath==1.3.0
msgpack==1.0.8
networkx==3.2.1
ninja==1.11.1.1
numba==0.59.1
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-ml-py==12.560.30
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.6.77
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==24.0
pandas==1.3.5
pathlib2==2.3.7.post1
pathtools==0.1.2
peft==0.11.1
pillow==10.3.0
platformdirs==4.2.2
pluggy==1.5.0
ply==3.11
pooch==1.8.1
portalocker==2.8.2
prettytable==3.10.0
promise==2.3
protobuf==3.20.3
psutil==5.9.8
ptflops==0.7.3
py==1.11.0
py-cpuinfo==9.0.0
py-spy==0.3.14
pyarrow==6.0.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pybind11==2.12.0
pycparser==2.22
pycryptodomex==3.20.0
pycurl==7.43.0.6
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.18.0
PyGObject==3.38.0
PyJWT==2.8.0
pyOpenSSL==24.2.1
pyparsing==3.1.2
pyPEG2==2.15.2
pytest==6.2.5
python-apt==2.2.1
python-consul==1.1.0
python-dateutil==2.9.0.post0
python-engineio==4.9.1
python-etcd==0.4.5
python-socketio==5.11.4
pytz==2022.5
PyYAML==6.0.1
qwen-vl-utils==0.0.8
redis==4.5.5
regex==2024.5.15
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.7.1
rsa==4.9
safetensors==0.4.3
schedule==1.2.2
scikit-learn==1.5.0
scipy==1.13.1
seaborn==0.13.2
sentencepiece==0.2.0
sentry-sdk==2.14.0
setproctitle==1.3.3
shortuuid==1.0.13
simple-websocket==1.0.0
six==1.16.0
smmap==5.0.1
soundfile==0.12.1
soxr==0.3.7
sympy==1.12.1
tabulate==0.9.0
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow-cpu==2.8.0
tensorflow-io==0.25.0
tensorflow-io-gcs-filesystem==0.25.0
termcolor==2.4.0
terminaltables==3.1.10
text-cutter @ https://luban-source.byted.org/repository/scm/search.nlp.libcut_py_2.3.0.48.tar.gz#sha256=46b14d7087fcc5bd3b5c5d0594849c45968df4d7b18d11b8cae02675b2097264
tf-estimator-nightly==2.8.0.dev2021122109
threadpoolctl==3.5.0
thrift2pyi==1.0.3
thriftpy2==0.5.0
tiktoken==0.8.0
timm==0.6.7
tokenizers==0.20.1
toml==0.10.2
tomli==2.0.1
torch==2.4.1
torchaudio @ file:///tmp/torchaudio/torchaudio-2.1.0rc4%2Bcu122-cp39-cp39-linux_x86_64.whl#sha256=ff4c4c237e7734445f4b30366518a5ccc02b950171701b86ec6661132446b584
torchlibrosa==0.1.0
torchvision @ file:///tmp/torchvision/torchvision-0.16.0rc4%2Bcu122-cp39-cp39-linux_x86_64.whl#sha256=791f97b3073a075d49930d0d447fde29ff42d925dec8464182b2369094616101
tox==3.28.0
tqdm==4.48.2
transformers==4.45.1
transformers-stream-generator==0.0.4
triton==3.0.0
typing_extensions==4.12.1
unattended-upgrades==0.1
urllib3==1.26.18
virtualenv==20.26.5
watchdog==5.0.2
wcwidth==0.2.13
websocket-client==1.8.0
Werkzeug==3.0.3
wrapt==1.16.0
wsproto==1.2.0
yacs==0.1.8
yapf==0.40.2
zipp==3.19.1
zope.event==5.0
zope.interface==6.4.post2
zstandard==0.19.0
